Move hbaseenrichmentwriter to the writer project
