+++ metron-platform/metron-writer/pom.xml
@@ -218,9 +218,14 @@
             </exclusions>
             <scope>provided</scope>
         </dependency>
-         <dependency>
+        <dependency>
             <groupId>org.apache.metron</groupId>
             <artifactId>metron-common</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+         <dependency>
+            <groupId>org.apache.metron</groupId>
+            <artifactId>metron-enrichment-common</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
@@ -228,6 +233,13 @@
             <artifactId>metron-test-utilities</artifactId>
             <version>${project.parent.version}</version>
             <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.metron</groupId>
+            <artifactId>metron-hbase</artifactId>
+            <version>${project.parent.version}</version>
+            <scope>test</scope>
+            <type>test-jar</type>
         </dependency>
         <dependency>
             <groupId>org.apache.metron</groupId>
